+++ src/js/Lists/ListItem.js
@@ -208,6 +208,30 @@ export default class ListItem extends PureComponent {
      * prop is `true`.
      */
     activeClassName: PropTypes.string,
+
+    /**
+     * Any additional props you would like to supply to the surrounding `<li>` tag for the `ListItem`.
+     * By default, all props will be provided to the inner `AccessibleFakeButton`. If the `passPropsToItem`
+     * prop is enabled, the remaining props will be provided to the `<lI>` tag instead and this prop
+     * is probably useless.
+     */
+    itemProps: PropTypes.object,
+
+    /**
+     * Any additional props you would like to add to the inner `AccessibleFakeButton`. By default, all the
+     * remaining props will be provided to the `AccessibleFakeButton`, so this prop is probably useless.
+     * Enabling the `passPropsToItem` prop will change the default behavior so that the remaining props
+     * are provided to the surrounding `<li>` node instead and this prop becomes usefull.
+     */
+    tileProps: PropTypes.object,
+
+    /**
+     * All the remaining props should be passed to the surrounding `<li>` node instead of the `AccessibleFakeButton`.
+     *
+     * > NOTE: This will most likely become the default in the next *major* release. Migration warnings will be added
+     * if that is the case.
+     */
+    passPropsToItem: PropTypes.bool,
     initiallyOpen: deprecated(PropTypes.bool, 'Use `defaultOpen` instead'),
   };
 
@@ -384,6 +408,10 @@ export default class ListItem extends PureComponent {
       activeClassName,
       expanderIconChildren,
       expanderIconClassName,
+      itemProps,
+      tileProps,
+      passPropsToItem,
+      component,
       ...props
     } = this.props;
     delete props.isOpen;
@@ -434,6 +462,8 @@ export default class ListItem extends PureComponent {
 
     return (
       <li
+        {...itemProps}
+        {...(passPropsToItem ? props : undefined)}
         style={style}
         className={cn('md-list-item', {
           'md-list-item--nested-container': nestedItems,
@@ -442,7 +472,9 @@ export default class ListItem extends PureComponent {
       >
         {prependNested ? nestedList : null}
         <AccessibleFakeInkedButton
-          {...props}
+          {...tileProps}
+          {...(passPropsToItem ? undefined : props)}
+          component={component}
           __SUPER_SECRET_REF__={this._setTile}
           key="tile"
           onClick={this._handleClick}