+++ src/js/TextFields/TextArea.js
@@ -1,6 +1,7 @@
 import React, { PureComponent } from 'react';
 import PropTypes from 'prop-types';
 import cn from 'classnames';
+import ResizeObserver from '../Helpers/ResizeObserver';
 
 /**
  * The `TextArea` component is used to allow a dynamic height for the
@@ -18,7 +19,6 @@ export default class TextArea extends PureComponent {
     defaultValue: PropTypes.string,
     floatingLabel: PropTypes.bool,
     value: PropTypes.string,
-    onHeightChange: PropTypes.func,
     block: PropTypes.bool,
     label: PropTypes.node,
   };
@@ -30,29 +30,26 @@ export default class TextArea extends PureComponent {
     this.focus = this.focus.bind(this);
     this.getField = this.getField.bind(this);
     this.getValue = this.getValue.bind(this);
+    this._setMask = this._setMask.bind(this);
+    this._setField = this._setField.bind(this);
     this._handleChange = this._handleChange.bind(this);
     this._handleResize = this._handleResize.bind(this);
     this._syncHeightWithMask = this._syncHeightWithMask.bind(this);
   }
 
   componentDidMount() {
-    this._rowHeight = this._calcRowHeight(this._field, this.props);
+    this._rowHeight = this._calcRowHeight();
     this._syncHeightWithMask();
-    window.addEventListener('resize', this._handleResize);
   }
 
   componentWillReceiveProps(nextProps) {
     if (this.props.rows !== nextProps.rows) {
-      this._rowHeight = this._calcRowHeight(this._field, this.props);
+      this._rowHeight = this._calcRowHeight(nextProps);
     }
 
     if (this.props.value !== nextProps.value || this.props.maxRows !== nextProps.maxRows) {
       this._syncHeightWithMask(nextProps.value);
     }
-  }
-
-  componentWillUnmount() {
-    window.removeEventListener('resize', this._handleResize);
   }
 
   getField() {
@@ -71,6 +68,18 @@ export default class TextArea extends PureComponent {
     this._field.blur();
   }
 
+  _calcRowHeight = ({ rows } = this.props) => {
+    if (!this._field) {
+      return 19;
+    }
+
+    const height = this._field.style.height;
+    this._field.style.height = 'auto';
+    const rowHeight = this._field.offsetHeight / rows;
+    this._field.style.height = height;
+    return rowHeight;
+  };
+
   _setMask(mask) {
     this._mask = mask;
   }
@@ -79,12 +88,8 @@ export default class TextArea extends PureComponent {
     this._field = field;
   }
 
-  _calcRowHeight(field, props) {
-    return field.offsetHeight / props.rows;
-  }
-
   _handleResize() {
-    this._rowHeight = this._calcRowHeight(this._field, this.props);
+    this._rowHeight = this._calcRowHeight();
     this._syncHeightWithMask();
   }
 
@@ -104,13 +109,6 @@ export default class TextArea extends PureComponent {
     }
 
     height = Math.max(this._rowHeight * rows, height);
-
-    if (this.props.onHeightChange) {
-      // For some reason the md-text-field-multiline-container is 5px
-      // larger than the textareas.. So just add 5 here and on the inline style
-      this.props.onHeightChange(height + 5);
-    }
-
     this.setState({ height });
   }
 
@@ -136,7 +134,6 @@ export default class TextArea extends PureComponent {
     } = this.props;
     delete props.maxRows;
     delete props.onChange;
-    delete props.onHeightChange;
 
     return (
       <div
@@ -146,8 +143,9 @@ export default class TextArea extends PureComponent {
           'md-text-field--floating-margin': label && !block,
         })}
       >
+        <ResizeObserver watchWidth onResize={this._handleResize} />
         <textarea
-          ref={mask => { this._mask = mask; }}
+          ref={this._setMask}
           className={cn(className, 'md-text-field--multiline-mask')}
           readOnly
           rows={props.rows}
@@ -159,7 +157,7 @@ export default class TextArea extends PureComponent {
         />
         <textarea
           {...props}
-          ref={field => { this._field = field; }}
+          ref={this._setField}
           style={Object.assign({}, style, { height })}
           className={className}
           defaultValue={defaultValue}